Changes testing framework

+++ lib/kublog/notification/email.rb
@@ -55,16 +55,18 @@ module Kublog
         def build_email_body
           if self.email_body.blank? && !Sanitize.clean(self.body).blank?
             template_path = "app/views/kublog/post_mailer/new_post.liquid.html.erb"
-            template = File.open(File.join(Rails.root, template_path))
+            if File.exists?(File.join(Rails.root, template_path))
+              template = File.open(File.join(Rails.root, template_path))
+            else
+              template = File.open(File.join(Engine.root, template_path))
+            end
             read_erb_template(template)
           end
         end
-
         def read_erb_template(template)
           post = self
           self.email_body = ERB.new(template.read).result(binding)
         end
-      
       end
     
       # Queues Process on DJ
